move classes into separate header and source files
