+++ spot_driver_cpp/include/spot_driver_cpp/interfaces/publisher_interface_base.hpp
@@ -1,0 +1,20 @@
+// Copyright (c) 2023 Boston Dynamics AI Institute LLC. All rights reserved.
+
+#pragma once
+
+#include <spot_driver_cpp/spot_image_sources.hpp>
+
+namespace spot_ros2
+{
+/**
+ * @brief Defines an interface for a class that publishes image data to middleware.
+ */
+class PublisherInterfaceBase
+{
+public:
+  virtual ~PublisherInterfaceBase() {};
+
+  virtual void createPublishers(const std::vector<ImageSource>& image_sources) = 0;
+  virtual void publish(const std::map<ImageSource, ImageWithCameraInfo>& images) = 0;
+};
+}
