+++ spot_driver_cpp/src/interfaces/rclcpp_wall_timer_interface.cpp
@@ -1,0 +1,21 @@
+// Copyright (c) 2023 Boston Dynamics AI Institute LLC. All rights reserved.
+
+#include <spot_driver_cpp/interfaces/rclcpp_wall_timer_interface.hpp>
+
+namespace spot_ros2
+{
+RclcppWallTimerInterface::RclcppWallTimerInterface(const std::shared_ptr<rclcpp::Node>& node)
+: node_{node}
+{
+}
+
+void RclcppWallTimerInterface::setTimer(const std::chrono::duration<double>& period, const std::function<void()>& callback)
+{
+    timer_ = node_->create_wall_timer(period, callback);
+}
+
+void RclcppWallTimerInterface::clearTimer()
+{
+    timer_.reset();
+}
+}
