+++ spot_driver_cpp/src/spot_image_publisher_component.cpp
@@ -1,0 +1,7 @@
+// Copyright (c) 2023 Boston Dynamics AI Institute LLC. All rights reserved.
+
+#include <rclcpp/node_options.hpp>
+#include <rclcpp_components/register_node_macro.hpp>
+#include <spot_driver_cpp/spot_image_publisher_node.hpp>
+
+RCLCPP_COMPONENTS_REGISTER_NODE(spot_ros2::SpotImagePublisherNode)
