+++ spot_driver_cpp/src/spot_image_publisher_node.cpp
@@ -1,19 +1,18 @@
 // Copyright (c) 2023 Boston Dynamics AI Institute LLC. All rights reserved.
 
-#include <rclcpp/executors.hpp>
-#include <rclcpp/rclcpp.hpp>
-#include <rclcpp/utilities.hpp>
-#include <spot_driver_cpp/spot_image_publisher.hpp>
+#include <spot_driver_cpp/spot_image_publisher_node.hpp>
 
-#include <memory>
-
-int main(int argc, char* argv[])
+namespace spot_ros2
 {
-  rclcpp::init(argc, argv);
+SpotImagePublisherNode::SpotImagePublisherNode(const rclcpp::NodeOptions& node_options)
+: node_{ std::make_shared<rclcpp::Node>( "image_publisher" , node_options) }
+, internal_{ SpotImagePublisher{ node_ } }
+{
+    internal_.initialize();
+}
 
-  const auto node = std::make_shared<spot_ros2::SpotImagePublisherNode>();
-
-  rclcpp::spin(node->get_node_base_interface());
-
-  return 0;
+std::shared_ptr<rclcpp::node_interfaces::NodeBaseInterface> SpotImagePublisherNode::get_node_base_interface()
+{
+    return node_->get_node_base_interface();
+}
 }
