+++ spot_driver_cpp/src/spot_image_publisher_node_main.cpp
@@ -1,0 +1,18 @@
+// Copyright (c) 2023 Boston Dynamics AI Institute LLC. All rights reserved.
+
+#include <rclcpp/executors.hpp>
+#include <rclcpp/utilities.hpp>
+#include <spot_driver_cpp/spot_image_publisher_node.hpp>
+
+#include <memory>
+
+int main(int argc, char* argv[])
+{
+  rclcpp::init(argc, argv);
+
+  const auto node = std::make_shared<spot_ros2::SpotImagePublisherNode>();
+
+  rclcpp::spin(node->get_node_base_interface());
+
+  return 0;
+}
